Fix zaps compile issue where proof contains reference into store
The problem: zaps creates a proof and passes in a reference to the
store (which it created). The proof contains a reference to the
store. But the compiler doesn't allow that because the store will be
destroyed after the zaps function finishes.

H/T @porcuquine for finding where this was happening.